# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bidragskoll.se API",
    "description": "Funding discovery, eligibility, application and case-management platform.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "profileId",
                  "title"
                ],
                "properties": {
                  "profileId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "intent": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "facts": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "totalBudgetMinor": {
                    "type": "integer",
                    "minimum": 0,
                    "nullable": true
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "startsOn": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "endsOn": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/projects/{id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "profileId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  },
                  "intent": {
                    "type": "string",
                    "maxLength": 5000
                  },
                  "facts": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "totalBudgetMinor": {
                    "type": "integer",
                    "minimum": 0,
                    "nullable": true
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  },
                  "startsOn": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "endsOn": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/projects/{id}/matches": {
      "post": {
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/projects/{id}/funding-stack": {
      "post": {
        "tags": [
          "matches"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownContributionMinor": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}